Pyrosphere drowning
-Pyrosphere is now crippled when touching water. Can no longer inflict On Fire!, spawn its projectile or produce dust. It's also drawn without the orange glow.
-Unchained Pyrosphere has the same nerfs.

+++ Projectiles/Melee/Flails/PyrosphereBall.cs
@@ -61,6 +61,8 @@ namespace MultidimensionMod.Projectiles.Melee.Flails
         // This AI code was adapted from vanilla code: Terraria.Projectile.AI_015_Flails() 
         public override void AI()
         {
+            if (Projectile.wet)
+                drowning = true;
             Player player = Main.player[Projectile.owner];
             // Kill the projectile if the player dies or gets crowd controlled
             if (!player.active || player.dead || player.noItems || player.CCed || Vector2.Distance(Projectile.Center, player.Center) > 900f)
@@ -157,9 +159,20 @@ namespace MultidimensionMod.Projectiles.Melee.Flails
                         }
                         if (shouldSwitchToRetracting)
                         {
-                            if (Main.myPlayer == Projectile.owner)
-                                Projectile.NewProjectile(Projectile.GetSource_FromAI(), Projectile.Center, Projectile.velocity, ModContent.ProjectileType<PyrosphereUnchained>(), Projectile.damage / 2, Projectile.knockBack, Main.myPlayer);
-                            Projectile.Kill();
+                            if (!drowning)
+                            {
+                                if (Main.myPlayer == Projectile.owner)
+                                    Projectile.NewProjectile(Projectile.GetSource_FromAI(), Projectile.Center, Projectile.velocity, ModContent.ProjectileType<PyrosphereUnchained>(), Projectile.damage / 2, Projectile.knockBack, Main.myPlayer);
+                                Projectile.Kill();
+                            }
+                            else
+                            {
+                                CurrentAIState = AIState.Retracting;
+                                StateTimer = 0f;
+                                Projectile.netUpdate = true;
+                                Projectile.velocity *= 0.3f;
+                            }
+                                
                         }
                         player.ChangeDir((player.Center.X < Projectile.Center.X).ToDirectionInt());
                         Projectile.localNPCHitCooldown = movingHitCooldown;
@@ -328,7 +341,7 @@ namespace MultidimensionMod.Projectiles.Melee.Flails
             if (doFastThrowDust)
                 dustRate = 1;
 
-            if (Main.rand.NextBool(dustRate))
+            if (Main.rand.NextBool(dustRate) && !drowning)
                 Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, DustID.Torch, 0f, 0f, 150, default(Color), 1.3f);
         }
 
@@ -545,20 +558,30 @@ namespace MultidimensionMod.Projectiles.Melee.Flails
                 {
                     Vector2 drawPos = Projectile.oldPos[k] - Main.screenPosition + drawOrigin + new Vector2(0f, Projectile.gfxOffY);
                     Color color = Color.White * ((float)(Projectile.oldPos.Length - k) / (float)Projectile.oldPos.Length);
-                    Main.spriteBatch.Draw(projectileTexture, drawPos, null, color, Projectile.rotation, drawOrigin, Projectile.scale - k / (float)Projectile.oldPos.Length / 3, spriteEffects, 0f);
+                    if (!drowning)
+                        Main.spriteBatch.Draw(projectileTexture, drawPos, null, color, Projectile.rotation, drawOrigin, Projectile.scale - k / (float)Projectile.oldPos.Length / 3, spriteEffects, 0f);
                 }
             }
             Texture2D texture = TextureAssets.Projectile[Projectile.type].Value;
+            Texture2D textureOff = ModContent.Request<Texture2D>("MultidimensionMod/Projectiles/Melee/Flails/PyrosphereBallExtinguished").Value;
             Vector2 drawOrigine = new(texture.Width / 2, texture.Height / 2);
             var effects = Projectile.spriteDirection == 1 ? SpriteEffects.None : SpriteEffects.FlipHorizontally;
 
-            Main.EntitySpriteDraw(texture, Projectile.Center - Main.screenPosition, null, Projectile.GetAlpha(Color.White), Projectile.rotation, drawOrigine, Projectile.scale, effects, 0);
-            return true;
+            if (drowning)
+                Main.EntitySpriteDraw(textureOff, Projectile.Center - Main.screenPosition, null, Projectile.GetAlpha(lightColor), Projectile.rotation, drawOrigine, Projectile.scale, effects, 0);
+            else
+                Main.EntitySpriteDraw(texture, Projectile.Center - Main.screenPosition, null, Projectile.GetAlpha(Color.White), Projectile.rotation, drawOrigine, Projectile.scale, effects, 0);
+            if (drowning)
+                return false;
+            else
+                return true;
         }
 
+        public bool drowning = false;
         public override void OnHitNPC(NPC target, NPC.HitInfo hit, int damageDone)
         {
-            target.AddBuff(BuffID.OnFire, 180);
+            if (drowning)
+                target.AddBuff(BuffID.OnFire, 180);
         }
     }
 }
